Fix: Change SequenceNumber from int to long for large values
- Changed Workstation.SequenceNumber from int? to long?
- Fixes JSON parsing error for large sequence numbers
- Int32 max (2,147,483,647) was too small for values like 987,654,345,678
- Long max (9,223,372,036,854,775,807) can handle much larger numbers

Resolves: "JSON value could not be converted to System.Nullable`1[System.Int32]"

+++ PubSubApp/Program.cs
@@ -856,7 +856,7 @@ partial class Program
         public string? Type { get; set; }
 
         [JsonPropertyName("sequenceNumber")]
-        public int? SequenceNumber { get; set; }
+        public long? SequenceNumber { get; set; }
     }
 
     public class References
